Use DefragmentAllDisks to raise faults
When a VM is powered on it can't DefragmentAllDisks so this
method can be used to always raise an error. These kinds of
tests are important because we can use them to test the XML
parser's handling of failure modes.

It also validates that the pyvmomi framework can handle true faults.